# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuários",
                "description": "Retorna uma lista de usuários cadastrados, com suporte a paginação e filtros.",
                "operationId": "dcc6d75003742c424fc8c95bf20ad811",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de itens por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar usuários pelo nome.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filtrar usuários pelo email.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "john.doe@example.com"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por status ativo (true) ou inativo (false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Criar um novo usuário",
                "description": "Adiciona um novo usuário ao banco de dados.",
                "operationId": "1932e13e89a3b0a0a1a75242f421b8f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário.",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email do usuário.",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário.",
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "active": {
                                        "description": "Status do usuário (ativo ou inativo).",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "roles": {
                                        "description": "Lista de IDs dos papéis associados ao usuário.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obter um usuário",
                "description": "Retorna os detalhes de um usuário específico.",
                "operationId": "fc21138408664b1ddc707c7fb520e1c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único do usuário.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualizar um usuário",
                "description": "Atualiza as informações de um usuário existente.",
                "operationId": "9344f2a75f244b8c62f5cadba56d8076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único do usuário.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome atualizado do usuário.",
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "description": "Email atualizado do usuário.",
                                        "type": "string",
                                        "example": "jane.doe@example.com"
                                    },
                                    "active": {
                                        "description": "Status atualizado do usuário.",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "roles": {
                                        "description": "Lista de IDs atualizados dos papéis associados ao usuário.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Usuário atualizado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Excluir um usuário",
                "description": "Remove um usuário do banco de dados.",
                "operationId": "7600bb0110feca273f95f048feef8955",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único do usuário.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}